move print and nav to app actions
