+++ web/src/actions/app/index.js
@@ -1,2 +1,4 @@
 export * from './apd';
+export * from './navigation';
+export * from './print';
 export * from './symbols';
