+++ web/src/actions/app/print.js
@@ -1,6 +1,8 @@
-export const PRINT_APD = Symbol('print : apd');
+import { PRINT_APD } from './symbols';
 
 export const printApd = ({ global = window } = {}) => dispatch => {
   dispatch({ type: PRINT_APD });
   global.print();
 };
+
+export default { printApd };
