+++ web/src/actions/app/symbols.js
@@ -10,6 +10,10 @@ export const FETCH_ALL_APDS_FAILURE = Symbol('fetch all apds failure');
 export const FETCH_ALL_APDS_REQUEST = Symbol('fetch all apds request');
 export const FETCH_ALL_APDS_SUCCESS = Symbol('fetch all apds success');
 
+export const NAVIGATION_SCROLL_TO_WAYPOINT = Symbol('scroll to waypoint');
+
+export const PRINT_APD = Symbol('print apd');
+
 export const SAVE_APD_FAILURE = Symbol('save apd failure');
 export const SAVE_APD_REQUEST = Symbol('save apd request');
 export const SAVE_APD_SUCCESS = Symbol('save apd success');
