+++ web/src/containers/ConnectedWaypoint.js
@@ -3,7 +3,7 @@ import React, { Component, Fragment } from 'react';
 import { connect } from 'react-redux';
 import { Waypoint } from 'react-waypoint';
 
-import { scrollTo } from '../actions/navigation';
+import { scrollTo } from '../actions/app';
 
 class ConnectedWaypoint extends Component {
   hitWaypoint = id => () => {
