+++ web/src/containers/viewOnly/Export.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import { Button } from '@cmsgov/design-system-core';
 import { FileDownload } from '../../components/Icons';
-import { printApd } from '../../actions/print';
+import { printApd } from '../../actions/app';
 
 const ExportInstructions = ({ printApd: print }) => {
   return (
